feat: Route authentication implemented
All Routes except the login and register routes are now protected by the JWT authentication middleware. The JWT is sent in the request header as a bearer token. The JWT is verified by the middleware and the control is passed to the route handler if the token is valid. If the token is invalid, the middleware returns a 401 status code and a message.